mainform: circleview: Split and decouple some elements
This moves circleview rendering and mouse updating into the circleView class.
The goal of this is to start to decouple circleView from needing direct
access to GUI elements, and to also help remove some non-GUI logic from
MainForm.

Most of the logic is left untouched. There are some small changes, mostly
with the if statements when checking if we need to pay attention to certain
mouse movement, but the underlying behavior is left as-is.